Update fourmolu to allow `single-constraint-parens: never`
The version of fourmolu included in Stack LTS 21.3 included an ormolu
change that always put parens around single constraints.

The option to configure this was added in the next version, and the most
recent version, fixed a problem with disabling this, so we have moved to
the most recent fourmolu now.